move test data to utils folder
