+++ src/components/EventCard/EventCard.spec.js
@@ -1,28 +1,9 @@
 import '@testing-library/jest-dom';
 import {render, screen} from '@testing-library/react';
 
-import EventCard from './EventCard';
+import dummyEvent from '../../utils/data';
 
-const dummyEvent = {
-	pictures: {
-		eventPicture: 'https://www.meetandeat-berlin.de/img/logo.svg',
-		coverPicture: '',
-	},
-	adress: {
-		street: '',
-		city: '',
-	},
-	private: false,
-	categories: [],
-	_id: '62b185e11a4caaa125207cab',
-	eventName: 'Event1',
-	eventPicture: '',
-	coverPicture: '',
-	participates: [],
-	desc: 'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam',
-	date: 'THU 30. Jun',
-	location: [],
-};
+import EventCard from './EventCard';
 
 describe('Event Card Component', () => {
 	it('should render correctly props', () => {
