+++ src/utils/data.js
@@ -1,0 +1,22 @@
+const dummyEvent = {
+	pictures: {
+		eventPicture: 'https://www.meetandeat-berlin.de/img/logo.svg',
+		coverPicture: '',
+	},
+	adress: {
+		street: '',
+		city: '',
+	},
+	private: false,
+	categories: [],
+	_id: '62b185e11a4caaa125207cab',
+	eventName: 'Event1',
+	eventPicture: '',
+	coverPicture: '',
+	participates: [],
+	desc: 'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam',
+	date: 'THU 30. Jun',
+	location: [],
+};
+
+export default dummyEvent;
